Add a ability to ignore based on anchor's InnerHtml
Added a ability to ignore based on anchor's InnerHtml
Added a note to console when you exceed the numbers of pages found on a site
Added 17buddies.rocks.  Does not current download, beacuse of CAPTCHA, but used to get a list of files, so I could compare against what I already have.